backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
alembic
python-dotenv
pytest
pydantic
